fixing some technical errors

+++ ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: oakhmouc <oakhmouc@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/05 00:25:53 by oakhmouc          #+#    #+#             */
-/*   Updated: 2024/12/09 21:27:53 by oakhmouc         ###   ########.fr       */
+/*   Updated: 2024/12/10 13:49:11 by oakhmouc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ int	check_p(void *p)
 	if (!p)
 		i = ft_putstr_fd("(nil)", 1);
 	else
-		i = ft_putadr((long unsigned int)p);
+		i = ft_putadr((unsigned long long)p);
 	return (i);
 }
 
@@ -54,7 +54,7 @@ int	print_var(va_list str, const char	*str1, int i)
 	else if (str1[i] == 'u')
 		ret = ft_putuns(va_arg(str, unsigned int), 1);
 	else if (str1[i] == 'p')
-		ret = check_p(va_arg(str, void*));
+		ret = check_p(va_arg(str, void *));
 	else if (str1[i] == '%')
 		ret = ft_putchar_fd('%', 1);
 	return (ret);
